+++ eslint.config.js
@@ -20,7 +20,6 @@ export default defineConfig([
     },
     rules: {
       'react-hooks/set-state-in-effect': 'off',
-      'react-hooks/exhaustive-deps': 'off',
     },
   },
 ])