+++ src/windows/LoginWindow/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { ConfigProvider } from 'antd';
 import 'antd/dist/reset.css';
-import { antdProviderProps } from '../../shared/styles/antd-theme';
+import { antdProviderProps } from '@/shared/styles/antd-theme';
 import App from './App';
 import './index.scss';